OtherProjects/new_crm/requirements.txt
APScheduler==3.10.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect2==0.7.0
dataclasses-json==0.5.8
Django==3.2
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.12.4
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.53.1
html5lib==1.1
idna==3.10
lxml==5.3.0
marshmallow==3.22.0
marshmallow-enum==1.5.1
mypy_extensions==1.1.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
num2words==0.5.13
numpy==1.24.4
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.0.3
phonepe-sdk==1.1.0
pillow==10.4.0
protobuf==3.20.3
pycparser==2.22
pycryptodome==3.18.0
pydyf==0.10.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.5
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==1.26.12
weasyprint==61.2
webencodings==0.5.1
xhtml2pdf==0.2.16
xlrd==2.0.1
XlsxWriter==3.2.0
zopfli==0.2.3
